// src/client/src/models/Activity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * People in Need - Springboard Service
 * This backend is meant to be a simple CRUD service for supporting  collaboration between organizations and the beneficiaries supported by them.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { ActivityStatus } from './ActivityStatus';
import {
    ActivityStatusFromJSON,
    ActivityStatusFromJSONTyped,
    ActivityStatusToJSON,
} from './ActivityStatus';

/**
 * 
 * @export
 * @interface Activity
 */
export interface Activity {
    /**
     * 
     * @type {string}
     * @memberof Activity
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof Activity
     */
    organizationId?: string;
    /**
     * 
     * @type {string}
     * @memberof Activity
     */
    beneficiaryId?: string;
    /**
     * 
     * @type {string}
     * @memberof Activity
     */
    needId?: string;
    /**
     * 
     * @type {Date}
     * @memberof Activity
     */
    startDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof Activity
     */
    endDate?: Date;
    /**
     * 
     * @type {ActivityStatus}
     * @memberof Activity
     */
    status?: ActivityStatus;
    /**
     * 
     * @type {string}
     * @memberof Activity
     */
    comments?: string;
}

/**
 * Check if a given object implements the Activity interface.
 */
export function instanceOfActivity(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function ActivityFromJSON(json: any): Activity {
    return ActivityFromJSONTyped(json, false);
}

export function ActivityFromJSONTyped(json: any, ignoreDiscriminator: boolean): Activity {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'organizationId': !exists(json, 'organizationId') ? undefined : json['organizationId'],
        'beneficiaryId': !exists(json, 'beneficiaryId') ? undefined : json['beneficiaryId'],
        'needId': !exists(json, 'needId') ? undefined : json['needId'],
        'startDate': !exists(json, 'startDate') ? undefined : (new Date(json['startDate'])),
        'endDate': !exists(json, 'endDate') ? undefined : (new Date(json['endDate'])),
        'status': !exists(json, 'status') ? undefined : ActivityStatusFromJSON(json['status']),
        'comments': !exists(json, 'comments') ? undefined : json['comments'],
    };
}

export function ActivityToJSON(value?: Activity | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'organizationId': value.organizationId,
        'beneficiaryId': value.beneficiaryId,
        'needId': value.needId,
        'startDate': value.startDate === undefined ? undefined : (value.startDate.toISOString()),
        'endDate': value.endDate === undefined ? undefined : (value.endDate.toISOString()),
        'status': ActivityStatusToJSON(value.status),
        'comments': value.comments,
    };
}